repair: remove uses of maximum_token()
The argument to calculate_natural_endpoints() is passed through to
token_metadata_impl::first_token_index(), which searches a list for the next
(in ring order - potentially wrapping around) token greater or equal to the
argument. Because it's "or equal", the nonexisting maximum_token() in this
context is equivalent to minimum_token(), not greatest_token().