service: storage_proxy: remove uses of maximum_token()
The argument to get_live_sorted_endpoints() is passed through to
token_metadata_impl::first_token_index(), which searches a list for the next
(in ring order - potentially wrapping around) token greater or equal to the
argument. Because it's "or equal", the nonexisting maximum_token() in this
context is equivalent to minimum_token(), not greatest_token().

This patch includes a small behaviour change: if the range processed in
process_one_range is (+inf, +inf] (or [+inf, +inf], or any other variant of
this interval), then it will be omitted, while previously it would be added.

Other that the above, this patch should not change any behaviour.