dht: token-sharding: add maybe_token_for_next_shard()
token_for_next_shard() is one of the last users of maximum_token():
it returns it when there is no valid result. Let's use an optional to do that,
so we can get rid of maximum_token().

This patch adds a variant of token_for_next_shard() which returns an optional
instead of maximum_token(). Next patches migrate all users of
token_for_next_shard() to this variant, and then remove the old variant.