alternator: Move the whole starting code into a sched group
The controller won't have the database_config at hands to get
the sched group from. All other client services run the whole
controller start in the needed sched group, so prepare the
alternator controller for that.

To make it compile (and while-at-it) also move up the sharded
server and executor instances and the smp_service_group. All
of these will migrate onto the controller in the next patch.

( the indentation is deliberately left broken )

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>