mutation_writer: bucket_writer: consume: propagate _consume_fut if queue_reader_handle is_terminated
When the queue_reader_handle is terminated it was
either explicitly aborted or the reader was closed prematurely.

In this case _consume_fut should hold the root-cause error
(e.g. when compaction is stopped).  Return it instead
of trying to push the mutation fragment.
If no error is returned from _consume_fut, make to sure to
return either the queue_reader_handle error, if available,
or a generic error since the writer.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>