storage_proxy: specialize query_singular() for non-IN queries
query_singular() accepts a partition_range_vector, corresponding to an IN
query. But such queries are rare compared to single-partition queries,
and the generarlization is expensive: we need to map_reduce the partitions,
and collect results in a merger and trim them.

Specialize for the common case by having a fast path for non-IN queries.
The code was derived from the general case with everything related
to multiple partitions removed, and two continuations (once separated
by map_reduce) merged.

Some fat still remains, the code contains a few std::vector and
std::unordered_map manipulations, but they are left for later.

perf_simple_query --smp 1 --operations-per-shard 1000000 --task-quota-ms 10:

before: median 208080.16 tps ( 81.1 allocs/op,  15.1 tasks/op,   48871 insns/op)

after:  median 221415.47 tps ( 74.1 allocs/op,  12.1 tasks/op,   46667 insns/op)

So, a ~6% improvement in tps and 5% improvement in instructions per op.
Also large reduction in tasks and allocations.

The change of course increases the maintenance burden, but I believe
it is worth the speedup.